Move TimerInternalsFactory to runners/core-java
