+++ runners/core-java/src/main/java/org/apache/beam/runners/core/TimerInternalsFactory.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util.state;
+package org.apache.beam.runners.core;
 
 import java.io.Serializable;
 import org.apache.beam.sdk.annotations.Experimental;
